Preliminary updates - not functional code
