+++ config/model_config.yaml
@@ -1,0 +1,77 @@
+# Model Configuration for Supply Chain ML Project
+
+# Baseline Models
+
+logistic_regression:
+  penalty: "l2"
+  C: 1.0
+  max_iter: 1000
+  solver: "lbfgs"
+  random_state: 42
+
+decision_tree:
+  max_depth: 10
+  min_samples_split: 20
+  min_samples_leaf: 10
+  random_state: 42
+
+# Advanced Models
+
+random_forest:
+  n_estimators: 200
+  max_depth: 15
+  min_samples_split: 10
+  min_samples_leaf: 5
+  max_features: "sqrt"
+  n_jobs: -1
+  random_state: 42
+
+xgboost:
+  n_estimators: 300
+  learning_rate: 0.05
+  max_depth: 6
+  min_child_weight: 1
+  subsample: 0.8
+  colsample_bytree: 0.8
+  gamma: 0
+  reg_alpha: 0
+  reg_lambda: 1
+  random_state: 42
+
+lightgbm:
+  n_estimators: 300
+  learning_rate: 0.05
+  max_depth: 6
+  num_leaves: 31
+  subsample: 0.8
+  colsample_bytree: 0.8
+  reg_alpha: 0
+  reg_lambda: 1
+  random_state: 42
+
+# Hyperparameter Tuning
+
+hyperparameter_tuning:
+  method: "randomized_search"  # randomized_search, grid_search
+  n_iter: 50
+  cv: 5
+  scoring: "roc_auc"
+  n_jobs: -1
+
+  # Random Forest parameter grid
+  random_forest_param_grid:
+    n_estimators: [100, 200, 300]
+    max_depth: [10, 15, 20, 25]
+    min_samples_split: [5, 10, 20]
+    min_samples_leaf: [2, 5, 10]
+    max_features: ["sqrt", "log2"]
+
+  # XGBoost parameter grid
+  xgboost_param_grid:
+    n_estimators: [200, 300, 400]
+    learning_rate: [0.01, 0.05, 0.1]
+    max_depth: [4, 6, 8]
+    min_child_weight: [1, 3, 5]
+    subsample: [0.7, 0.8, 0.9]
+    colsample_bytree: [0.7, 0.8, 0.9]
+    gamma: [0, 0.1, 0.2]
