+++ requirements.txt
@@ -1,0 +1,31 @@
+# Data Processing
+pandas==2.3.3
+numpy==2.0.2
+boto3==1.41.5
+
+# SQL Database
+duckdb==1.1.0
+
+# Machine Learning
+scikit-learn==1.6.0
+xgboost==2.1.3
+lightgbm==4.5.0
+imbalanced-learn==0.12.0
+
+# Feature Engineering
+category_encoders==2.6.0
+
+# Visualization
+matplotlib==3.9.0
+seaborn==0.13.0
+
+# Model Explainability
+shap==0.46.0
+
+# Utilities
+pyyaml==6.0
+joblib==1.4.0
+
+# Development
+jupyter==1.1.1
+ipykernel==6.31.0
